Fix problem with pagination
If there are no posts in a topic or no comments on a post
the next page button should be disabled. This was not working due
to an incorrect equality check. The check should be that the page
is greater than or equal to the max page, not just equal to.